refactor: split storage facade to files
