+++ OptimoveSDK/Sources/Classes/Storage/OptimoveStorage.swift
@@ -1,0 +1,4 @@
+//  Copyright © 2023 Optimove. All rights reserved.
+
+/// Combined protocol for a convenince access to stored values and files.
+typealias OptimoveStorage = FileStorage & KeyValueStorage & StorageValue
